adk-agents/frontend_agent/requirements.txt
# Frontend Developer Agent - Dependencies
google-adk>=0.1.0
google-genai>=0.1.0
python-dotenv>=1.0.0
